% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTplates.R
\name{annotate_tRF}
\alias{annotate_tRF}
\title{Annotate a tRF sequence}
\usage{
annotate_tRF(sequence)
}
\arguments{
\item{sequence}{tRF sequence.}
}
\value{
tRF ID, type, whether exclusive to tRNA space, and tRNA sources of the tRF.
}
\description{
Obtain tRF ID, type, whether exclusive to tRNA space, and tRNA sources of the tRF with its sequence.
}
\examples{
sequence='TCCCTGGTGGTCTAGTGGTTAGGATTCGGC'
annotate_tRF(sequence)
}
\references{
Loher P, Telonis AG, Rigoutsos I. Sci Rep (2017) <doi: 10.1038/srep41184>
}
