% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTplates.R
\name{deseq}
\alias{deseq}
\title{Decode a license-plate}
\usage{
deseq(plate)
}
\arguments{
\item{plate}{The license plate being decoded.}
}
\value{
The sequence it decodes to.
}
\description{
Decode the license-plate using the lookup table.
}
\examples{
deseq('tRF-18-BS6PDFD2')
}
\references{
Pliatsika V, Loher P, Telonis AG, Rigoutsos I. Bioinformatics (2016) <doi: 10.1093/bioinformatics/btw194>
}
