% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTplates.R
\name{enseq}
\alias{enseq}
\title{Encode a sequence}
\usage{
enseq(sequence, prefix = "")
}
\arguments{
\item{sequence}{The sequence being encoded.}

\item{prefix}{The prefix to use for the license plate.}
}
\value{
The license plate it encodes to.
}
\description{
Encode the sequence into its corresponding license plate with given prefix (if given one).
}
\examples{
enseq('AACCGGGCAGAAGCACCA','tRF')
}
\references{
Pliatsika V, Loher P, Telonis AG, Rigoutsos I. Bioinformatics (2016) <doi: 10.1093/bioinformatics/btw194>
}
