% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTplates.R
\name{enseqs}
\alias{enseqs}
\title{Encode sequences}
\usage{
enseqs(sequences, prefix = "")
}
\arguments{
\item{sequences}{The sequences being encoded.}

\item{prefix}{The prefix to use for the license plate.}
}
\value{
The license plates they encode to.
}
\description{
Encode the sequences into their corresponding license plates with given prefix (if given one).
}
\examples{
seqs=c('AACCGGGCAGAAGCACCA','GAGCCCCAGTGGAACCACCA')
enseqs(seqs,'tRF')
}
\references{
Pliatsika V, Loher P, Telonis AG, Rigoutsos I. Bioinformatics (2016) <doi: 10.1093/bioinformatics/btw194>
}
