% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTplates.R
\name{exclusive}
\alias{exclusive}
\title{Determine whether the tRFs are exclusive to tRNA space}
\usage{
exclusive(sequences)
}
\arguments{
\item{sequences}{tRF sequences.}
}
\value{
Whether the tRFs are exclusive to tRNA space.
}
\description{
Determine whether the tRFs are exclusive to tRNA space with the tRF sequences.
}
\examples{
sequences=c('TCCCTGGTGGTCTAGTGGTTAGGATTCGGC','TCCCTGGTGGTCTAGTGGTTAGGATTCGGCG')
exclusive(sequences)
}
\references{
Loher P, Telonis AG, Rigoutsos I. Sci Rep (2017) <doi: 10.1038/srep41184>
}
