% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTplates.R
\name{source_tRNA}
\alias{source_tRNA}
\title{Obtain the tRNA source of a tRF}
\usage{
source_tRNA(sequence)
}
\arguments{
\item{sequence}{tRF sequence.}
}
\value{
Sources of the tRF.
}
\description{
Obtain the tRNA source of a tRF with the tRF sequence.
}
\examples{
sequence='TCCCTGGTGGTCTAGTGGTTAGGATTCGGC'
source_tRNA(sequence)
}
\references{
Loher P, Telonis AG, Rigoutsos I. Sci Rep (2017) <doi: 10.1038/srep41184>
}
