% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTplates.R
\name{source_tRNA2}
\alias{source_tRNA2}
\title{Obtain the tRNA sources of tRFs}
\usage{
source_tRNA2(sequences)
}
\arguments{
\item{sequences}{tRF sequences.}
}
\value{
Sources of the tRFs.
}
\description{
Obtain the tRNA sources of tRFs with the tRF sequences.
}
\examples{
sequences=c('TCCCTGGTGGTCTAGTGGTTAGGATTCGGC','TCCCTGGTGGTCTAGTGGCT','TCCCTGGTGGTCTAATGGTTA')
source_tRNA2(sequences)
}
\references{
Loher P, Telonis AG, Rigoutsos I. Sci Rep (2017) <doi: 10.1038/srep41184>
}
