% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINTplates.R
\name{tRFtype}
\alias{tRFtype}
\title{Obtain the type of tRFs}
\usage{
tRFtype(sequences)
}
\arguments{
\item{sequences}{tRF sequences.}
}
\value{
The type of tRFs.
}
\description{
Obtain the type of tRFs with the tRF sequences.
}
\examples{
sequences=c('TCCCTGGTGGTCTAGTGGTTAGGATTCGGC','AAAAATTTTGGTGCAACTCCAAATAAAA')
tRFtype(sequences)
}
\references{
Loher P, Telonis AG, Rigoutsos I. Sci Rep (2017) <doi: 10.1038/srep41184>
}
